Prevented delete scripts from running without sign-in

+++ Portal/php/deleteaccess.php
@@ -2,13 +2,19 @@
     include_once "database.php";
     $database = new Database();
 
+    include_once "sessions.php";
+
     $redirect = "../page/memberaccess.php";
 
-    if (isset($_GET['id'])) {
-        $id = $_GET['id'];
+    if (!isset($_SESSION['userid'])) {
+        $redirect = "../index.php";
+    } else {
+        if (isset($_GET['id'])) {
+            $id = $_GET['id'];
 
-        $database->removePrivilege($id);
-        $redirect = "../page/listaccess.php?id=$id";
+            $database->removePrivilege($id);
+            $redirect = "../page/listaccess.php?id=$id";
+        }
     }
 
     header("Location: $redirect");
