+++ Portal/php/deletereader.php
@@ -2,12 +2,18 @@
     include_once "database.php";
     $database = new Database();
 
+    include_once "sessions.php";
+
     $redirect = "../page/listreaders.php";
 
-    if (isset($_GET['id'])) {
-        $id = $_GET['id'];
+    if (!isset($_SESSION['userid'])) {
+        $redirect = "../index.php";
+    } else {
+        if (isset($_GET['id'])) {
+            $id = $_GET['id'];
 
-        $database->removeReader($id);
+            $database->removeReader($id);
+        }
     }
